pyproject.toml
[project]
name = "snowflake-labs-mcp"
version = "1.3.4"
description = "MCP server for Snowflake"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "Carlos Serrano" },
  { name = "Jason Summer" },
  { name = "Tyler White" },
]
dependencies = [
    "fastmcp>=2.8.1,<3.0.0",
    "mcp[cli]>=1.9.4,<2.0.0",
    "pydantic>=2.11.4,<3.0.0",
    "pyyaml>=6.0.2,<7.0.0",
    "requests>=2.32.3,<3.0.0",
    "snowflake-connector-python>=3.15.0,<4.0.0",
    "snowflake-core>=1.0.0,<2.0.0",
    "sqlglot>=27.8.0,<30.0.0",
]

[project.urls]
Repository = "https://github.com/Snowflake-Labs/mcp.git"
Issues = "https://github.com/Snowflake-Labs/mcp/issues"

[dependency-groups]
dev = [
    "fast-agent-mcp>=0.2.24,<2.0.0",
    "pre-commit>=4.2.0,<5.0.0",
    "pyright>=1.1.402,<2.0.0",
    "pytest>=8.3.5,<9.0.0",
    "python-dotenv>=1.1.0,<2.0.0",
    "ruff>=0.11.8,<1.0.0",
]

[build-system]
requires = ["hatchling>=1.0.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mcp_server_snowflake"]

[project.scripts]
mcp-server-snowflake = "mcp_server_snowflake:main"
snowflake-labs-mcp = "mcp_server_snowflake:main"
